requirements.txt
absl-py==1.0.0
amqp==5.1.0
anyio==3.5.0
asgiref==3.5.0
astunparse==1.6.3
bcrypt==3.2.0
billiard==3.6.4.0
cachetools==5.0.0
celery==5.2.3
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==36.0.1
cssselect==1.1.0
cssutils==2.4.0
cycler==0.11.0
dnspython==2.2.1
ecdsa==0.17.0
email-validator==1.1.3
emails==0.6
environs==9.5.0
fastapi==0.75.0
flatbuffers==2.0
fonttools==4.29.1
gast==0.5.3
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
h11==0.13.0
h5py==3.6.0
httptools==0.3.0
idna==3.3
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.2
kombu==5.2.4
libclang==13.0.0
lxml==4.8.0
Markdown==3.3.6
marshmallow==3.14.1
matplotlib==3.5.1
mkl-fft==1.3.1
mkl-random @ file:///home/builder/ci_310/mkl_random_1641843545607/work
mkl-service==2.4.0
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1639496814640/work
oauthlib==3.2.0
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.1
passlib==1.7.4
Pillow==9.0.1
premailer==3.10.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psycopg2-binary==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==1.9.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.3
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn @ file:///home/builder/ci_310/scikit-learn_1641557398975/work
scipy==1.8.0
seaborn==0.11.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio==1.2.0
SQLAlchemy==1.4.32
starlette==0.17.1
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.24.0
termcolor==1.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
typing_extensions==4.1.1
urllib3==1.26.8
uvicorn==0.17.5
uvloop==0.16.0
vine==5.0.0
watchgod==0.7
wcwidth==0.2.5
websockets==10.2
Werkzeug==2.0.3
wrapt==1.13.3
